rename: フォルダ名を変更
